\name{where_py}
\alias{where_py}
\title{Locate the path of the Python executable}
\description{
This function locates the path of the Python executable.
}
\usage{
where_py()
}
\value{
A character string specifying the location of the Python executable.
}

\keyword{Python}
\keyword{executable}
\keyword{location}